Add _IMAGE command and shared image rendering library
